refactor: reorganize MIDI handling and interface architecture
- Move interface definitions to dedicated Interface directory for better organization
- Separate MIDI processing concerns from main processor class
