+++ Source/Interface/ProgramChangeCallback.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class ProgramChangeCallback {
+public:
+    virtual ~ProgramChangeCallback() = default;
+    virtual void onProgramChange(int programNumber) = 0;
+};
