+++ Source/MidiHandler.cpp
@@ -1,0 +1,162 @@
+#include <MidiHandler.h>
+#include <Utils.h>
+#include <Engine/SynthEngine.h>
+#include "midiMap.h"
+#include "ObxdBank.h"
+#include "ParamaterManager.h"
+
+MidiHandler::MidiHandler(SynthEngine& s, MidiMap& b, ObxdBank& p, ParameterManager& pm, ProgramChangeCallback& pcCallback)
+    : synth(s), bindings(b), programs(p), paramManager(pm), programChangeCallback(pcCallback)
+{
+}
+
+
+void MidiHandler::prepareToPlay()
+{
+    nextMidi = std::make_unique<juce::MidiMessage>(0xF0);
+    midiMsg  = std::make_unique<juce::MidiMessage>(0xF0);
+}
+
+void MidiHandler::processMidiPerSample(juce::MidiBufferIterator* iter, const juce::MidiBuffer& midiBuffer, const int samplePos)
+{
+    while (getNextEvent(iter, midiBuffer, samplePos))
+    {
+        if (!midiMsg)
+            continue;
+
+        const auto size = midiMsg->getRawDataSize();
+        if (size < 1)
+            continue;
+
+        const auto* data = midiMsg->getRawData();
+        if (!data)
+            continue;
+
+        const auto status = data[0] & 0xF0;
+        if (status != 0x80 && status != 0x90 && status != 0xB0 &&
+            status != 0xC0 && status != 0xE0)
+            continue;
+
+        DBG("Valid Message: " << (int)midiMsg->getChannel() << " "
+            << (int)status << " "
+            << (size > 1 ? (int)data[1] : 0) << " "
+            << (size > 2 ? (int)data[2] : 0));
+
+        if (midiMsg->isNoteOn())
+        {
+            synth.procNoteOn(midiMsg->getNoteNumber(), midiMsg->getFloatVelocity());
+        }
+        else if (midiMsg->isNoteOff())
+        {
+            synth.procNoteOff(midiMsg->getNoteNumber());
+        }
+        if (midiMsg->isPitchWheel())
+        {
+            synth.procPitchWheel((midiMsg->getPitchWheelValue() - 8192) / 8192.0f);
+        }
+        if (midiMsg->isController() && midiMsg->getControllerNumber() == 1)
+        {
+            synth.procModWheel(midiMsg->getControllerValue() / 127.0f);
+        }
+        if (midiMsg->isSustainPedalOff() || midiMsg->isAllNotesOff() || midiMsg->isAllSoundOff())
+        {
+            synth.sustainOff();
+        }
+        if (midiMsg->isAllNotesOff())
+        {
+            synth.allNotesOff();
+        }
+        if (midiMsg->isAllSoundOff())
+        {
+            synth.allSoundOff();
+        }
+
+        DBG(" Message: " << midiMsg->getChannel() << " "
+            << " " << ((size > 2) ? midiMsg->getRawData()[2] : 0));
+
+        if (midiMsg->isProgramChange())  // xC0
+        {
+            programChangeCallback.onProgramChange(midiMsg->getProgramChangeNumber());
+        }
+        else if (midiMsg->isController()) // xB0
+        {
+            lastMovedController = midiMsg->getControllerNumber();
+            if (programs.currentProgramPtr->values[MIDILEARN] > 0.5f)
+            {
+                midiControlledParamSet = true;
+                bindings.updateCC(lastUsedParameter, lastMovedController);
+                juce::File midi_file = Utils::getMidiFolder().getChildFile("Custom.xml");
+                bindings.saveFile(midi_file);
+                currentMidiPath = midi_file.getFullPathName();
+
+                paramManager.setEngineParameterValue(synth, MIDILEARN, 0, true);
+                lastMovedController = 0;
+                lastUsedParameter = 0;
+                midiControlledParamSet = false;
+            }
+
+            if (bindings[lastMovedController] > 0)
+            {
+                midiControlledParamSet = true;
+                paramManager.setEngineParameterValue(synth, bindings[lastMovedController],
+                                        midiMsg->getControllerValue() / 127.0f, true);
+
+                paramManager.setEngineParameterValue(synth, MIDILEARN, 0, true);
+                lastMovedController = 0;
+                lastUsedParameter = 0;
+                midiControlledParamSet = false;
+            }
+        }
+    }
+}
+
+bool MidiHandler::getNextEvent(juce::MidiBufferIterator* iter, const juce::MidiBuffer& midiBuffer, const int samplePos)
+{
+	if (iter == nullptr)
+		return false;
+
+	if (*iter == midiBuffer.end())
+		return false;
+
+    if (const auto metadata = **iter; metadata.samplePosition >= samplePos && metadata.getMessage().getRawDataSize() > 0)
+	{
+		*midiMsg = metadata.getMessage();
+		midiEventPos = metadata.samplePosition;
+		++(*iter);
+		return true;
+	}
+    ++(*iter);
+    return getNextEvent(iter, midiBuffer, samplePos);
+}
+
+void MidiHandler::initMidi(){
+    //Documents > Obxd > MIDI > Default.xml
+    if (juce::File default_file = Utils::getMidiFolder().getChildFile("Default.xml"); !default_file.exists()){
+        bindings.saveFile(default_file);
+    }
+
+    const juce::File midi_config_file = Utils::getMidiFolder().getChildFile("Config.xml");
+    juce::XmlDocument xmlDoc (midi_config_file);
+
+    if (const std::unique_ptr<juce::XmlElement> ele_file = xmlDoc.getDocumentElementIfTagMatches("File")) {
+        const juce::String file_name = ele_file->getStringAttribute("name");
+        // Midi cc loading
+        if (juce::File midi_file = Utils::getMidiFolder().getChildFile(file_name); bindings.loadFile(midi_file)){
+            currentMidiPath = midi_file.getFullPathName();
+        } else {
+            if (juce::File xml = Utils::getMidiFolder().getChildFile("Default.xml"); bindings.loadFile(xml)){
+                currentMidiPath = xml.getFullPathName();
+            }
+        }
+    }
+}
+
+void MidiHandler::updateMidiConfig() const {
+    const juce::File midi_config_file = Utils::getMidiFolder().getChildFile("Config.xml");
+    juce::XmlDocument xmlDoc (midi_config_file);
+    if (const std::unique_ptr<juce::XmlElement> ele_file = xmlDoc.getDocumentElementIfTagMatches("File")) {
+        const juce::File f(currentMidiPath);
+        ele_file->setAttribute("name", f.getFileName());
+        ele_file->writeTo(midi_config_file.getFullPathName());
+    }
+}
